+++ pkg/filesystem/fs/config_test.go
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package nydus
+package fs
 
 import (
 	"testing"
