+++ pkg/filesystem/fs/testdata/invalid-bootstrap-file-size-133513.tar.gz
