+++ pkg/filesystem/fs/testdata/v5-bootstrap-file-size-736032.tar.gz
